Replace tag ids with display name
The ids for contributors, funders and publishers
are used to filter the search results page. However,
the tag that gets displayed should show the name of
the contributor/funder/publisher instead of the id
used for filtering.

Make a request to the api gateway bulk_agents endpoint
that makes a bulk request to SHARE to get record data.
Then replace the tag ids with their correct display names.